refactor02
move load_inpainting_pipline function to stereo_video_inpainting.py
