+++ pipelines/stereo_video_inpainting.py
@@ -731,3 +731,60 @@ def configure_attention_processors(pipeline: StableVideoDiffusionInpaintingPipel
             logger.warning(f"Failed to enable xFormers attention: {e}")
     if not attention_set:
         logger.info("Using default attention processor")
+
+def load_inpainting_pipeline(
+    pre_trained_path: str,
+    unet_path: str,
+    device: str = "cuda",
+    dtype: torch.dtype = torch.float16,
+    offload_type: str = "model"
+) -> StableVideoDiffusionInpaintingPipeline:
+    """
+    Loads the stable video diffusion inpainting pipeline components and returns the pipeline object.
+    """
+    logger.info("Loading pipeline components...")
+
+    image_encoder = CLIPVisionModelWithProjection.from_pretrained(
+        pre_trained_path,
+        subfolder="image_encoder",
+        variant="fp16",
+        torch_dtype=dtype,
+    )
+    vae = AutoencoderKLTemporalDecoder.from_pretrained(
+        pre_trained_path,
+        subfolder="vae",
+        variant="fp16",
+        torch_dtype=dtype,
+    )
+    unet = UNetSpatioTemporalConditionModel.from_pretrained(
+        unet_path,
+        subfolder="unet_diffusers",
+        low_cpu_mem_usage=True,
+        torch_dtype=dtype,
+    )
+
+    image_encoder.requires_grad_(False)
+    vae.requires_grad_(False)
+    unet.requires_grad_(False)
+
+    pipeline = StableVideoDiffusionInpaintingPipeline.from_pretrained(
+        pre_trained_path,
+        image_encoder=image_encoder,
+        vae=vae,
+        unet=unet,
+        torch_dtype=dtype,
+    ).to(device)
+
+    # NEW: Call the helper function to configure attention processors
+    configure_attention_processors(pipeline)
+    
+    if offload_type == "model":
+        pipeline.enable_model_cpu_offload()
+    elif offload_type == "sequential":
+        pipeline.enable_sequential_cpu_offload()
+    elif offload_type == "none":
+        pass  # No offloading
+    else:
+        raise ValueError("Invalid offload_type")
+
+    return pipeline
